firmware/requirements.txt
numpy==2.2.3
pandas==2.2.3
pydobot==1.3.2
pyserial==3.4
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
